+++ src/neuroglancer_annotation_ui/__init__.py
@@ -1,3 +1,3 @@
-from neuroglancer_annotation_ui.base import EasyViewer
-from neuroglancer_annotation_ui.utils import set_static_content_source, stop_ngl_server
+from neuroglancer_annotation_ui.easyviewer import EasyViewer, set_static_content_source, annotation
+from neuroglancer_annotation_ui import statebuilder
 __version__ = "1.0.0"
