+++ src/neuroglancer_annotation_ui/easyviewer/__init__.py
@@ -1,0 +1,3 @@
+from .base import EasyViewer
+from . import annotation
+from .utils import set_static_content_source
