+++ src/neuroglancer_annotation_ui/easyviewer/annotation.py
