+++ src/neuroglancer_annotation_ui/statebuilder/statebuilder.py
@@ -1,5 +1,5 @@
 from neuroglancer_annotation_ui import EasyViewer, annotation, set_static_content_source
-from neuroglancer_annotation_ui.utils import default_static_content_source
+from neuroglancer_annotation_ui.easyviewer.utils import default_static_content_source
 import pandas as pd
 import numpy as np
 from collections.abc import Collection
